Create new gradle task to lint c++ profiler code
The new gradle task is wrapping the existing clang-tidy cmake target
to make it easy to run linting on our code base. In the future I plan
to add additional linting steps (e.g. clang-format) and also plan to
run them on the build bot.

Additonally this CL contains a few fixes in the cmake side clang-tidy
integration to have better dependency handling.

Change-Id: I352619cc5e09769f44c1cab11b12ff9e111e5a7e

Former-commit-id: bb556f43fb6de30fdb28d7155c5228c0bcfd7301