Adds a fake ADB Server implementation to ddmlib.
This CL adds an extensible fake ADB Server to ddmlib.
The server can be used in both standalone or in a test.

The purpose of the server is to fake the ADB Server so that
consumers of ddmlib, as well as ddmlib, can be properly
tested.

The server is constructed such a way that a developer can
pick and chose what ADB commands to use/override. The
developer can change the behavior of the server to simulate
errors, device connection/disconnection, JDWP, or any other
internal workings of a hypothetical ADB Server.

Change-Id: I639838bd1b7c1fb07cdc8f5f50db094cbaa8a88d

Former-commit-id: ef1e24ee8818793194c09905dce4c92304b8718a