Improve dependency model and support for sub-modules
- Fix a bunch of issues in our model objects around equals and hashcode.
  This would trigger creating more objects than we needed when building
  the model passed to the IDE.

- Created LocalLibraryDependency which is for sub-modules. We don't extract
  their aar anymore and now point directly to the staged folder in their
  build output.

To do #2 I had to change where the staging is from nest folders representing
build types and flavors to a single folder with the full flavor name.
This is because the consumer has only access to the full flavor name and I
didn't want to have to recontruct the nested folder path.

Change-Id: I3f772cdb21826c3684d4210261bc7d63a3689131

Former-commit-id: b8fec59426fb4cae6b3f1c5ba0a66213117a0e9d