Not writing .knownpackages if there are no packages
The .knownpackages check was in RepoManagerImpl, which didn't really
make sense: what if, for example, the local packages were contained in a
db in a single file, or something? Additionally, LocalRepoLoader had to
expose some information that seemed like it should be internal. So
refactored the .knownpackages logic into LocalRepoLoaderImpl.

This made me notice that there was some extra complexity that wasn't
really necessary in RepoManagerImpl (running callbacks separately if
there's an ongoing task), and some logic that was incorrect (sometimes
we would reload the local repo when we shouldn't).

Also, added and enhanced some tests.

Fixes b.android.com/219655

Change-Id: I7130903049f24e0494ac7b20e03be8c19ea2a65f

Former-commit-id: 9bbc66664705687473d3a02d8798dc53704a3517