Not writing .knownpackages if there are no packages
Version two, now with fewer deadlocks!

The .knownpackages check was in RepoManagerImpl, which didn't really
make sense: what if, for example, the local packages were contained in a
db in a single file, or something? Additionally, LocalRepoLoader had to
expose some information that seemed like it should be internal. So
refactored the .knownpackages logic into LocalRepoLoaderImpl.

This made me notice that there was some extra complexity that wasn't
really necessary in RepoManagerImpl (running callbacks separately if
there's an ongoing task), and some logic that was incorrect (sometimes
we would reload the local repo when we shouldn't).

Also, added and enhanced some tests.

Fixes b.android.com/219655

Change-Id: I4bc4506098c979114f915d716ca539393fdbb282

Former-commit-id: f6bac3779a51801d63e8c73d344cfb65f09286b2