Disable most model test due to disable L2.
In 2.2-release the level 2 model query is disabled
since it's not set in stone what L2 means. However
a lot of tests expects things in the model based
on a L2 model since that's the default.

Since we have some level1 specific test that will
detect regressions, I'm disabling all failing tests
in the release branch only.

Change-Id: I27ba3e78962aa6ca67e6ea8ec12b9cf20646d88c

Former-commit-id: 89fbd1f048cd47bc22534632849e6488b3b8b75a