Fix dependencies leak to Jack annotation processing
If the class was loaded by the plugin's class loader, when
processing annotations, ECJ would use that version instead
of the one specified by classpath or annotation processor
paths. This was visible for Jack in-process. Now we mask all
the internal classes, and Jack has a limited visibility of
plugin's classes. This forces it to use the classes specified in
the classpath and APT path.

Bug: http://b.android.com/213471
Change-Id: Icc5dae271c4de46155873a726a6648b8a69a031c

Former-commit-id: a2e2bb7506c8c720447fc8c80b209750071d4d75