External native build rip out obsolete code copy to lib code
This code was stopgap for CMake since CMake didn't output a
stripped .so. We copied the unstripped .so to lib/. Since stripping
was implemented we no longer need this code.

Also address case where we were checking existence of .exe output
file even though build was only for .so files.

220727-change-abi

Change-Id: I1586c030eb4f935d5547ac84a0d60d104d2ddfe1
Former-commit-id: eb2ee7c58425fd01369944a25a23f7f68c65ae5b