Fix flavor merging
Overlay all the layers in order:
* defaultConfig
* dimensions (in order specified with "flavorDimensions")
* build type

This is how we used to do it in 2.1 and reflects the task names (values
are overridden/applied going from left to right, e.g.
assembleBlueBetaDebug applies the suffixes in the order
"blue.beta.debug"), even if it's inconsistent with our docs. I'll follow
up with the doc writers.

Bug: http://b.android.com/221641
Change-Id: I000e88e9f26f17b7deff9408f010e477ea8b94eb

Former-commit-id: d77e2ef01e766a1c5f8704360a2081c8c90c5a19