Remove the "minimal" rules for platforms
Instead exclude the biggest directories in the default rules. I defined
a macro to keep all the platform rules consistent. For now the tests
still pass, this may prove problematic if we ever get to Studio UI tests
that actually need the resources. Of course it would be best if Bazel
lifted the limit of number of symlinks by then.

Change-Id: I50c0c38b8e0821cc5ca80e9f0937b2b659e67dd5

Former-commit-id: 1854a0d9fb7ef0a7989dc8a1fe43ca351d71741d