Warn usage of getToken w/o refresh callback
The FirebaseInstanceId token is used by cloud messaging to send
notifications to individual devices. Application servers will store this
for contacting specific devices, and if an Android application reads it
with a getToken() call, but never observes changes to it via
onTokenRefresh, applications will not be able to deliver messages to
individual devices.

This change adds a "MissingFirebaseInstanceTokenRefresh" Android Lint
check to warn the user of this case.

NOTE: this change modifies the API of JavaPsiVisitor to allow a single
instance to visit both classes and method calls.

Change-Id: I4fdb429a93bdb6c4d3e626812f2085a4e05c909d

Former-commit-id: 59d91f448ad996331e70a3157f1e9cf8853f2b06