Build Info: add ability to serialize contents for logging
This CL keeps the contents of the build info file in memory so that it
can be provided later if the flight recorder needs it. We could possibly
just keep the reference to the path around, but that was a larger API
change, and the build-info is not a huge file anyway.

Change-Id: Ia6ee72afe2c597b6aa38c147b5e9852a6d4a8fd3
(cherry picked from commit b7c17d7d76f389f8aa0f8883dfdc76d36aa70e3a [formerly 63d60eb3eaffa94aedf7c70b63b23219d769cc68])

Former-commit-id: 9ab5d1e0577ae4d4106a560449b84621e207befe