Build Info: add ability to serialize contents for logging
This CL keeps the contents of the build info file in memory so that it
can be provided later if the flight recorder needs it. We could possibly
just keep the reference to the path around, but that was a larger API
change, and the build-info is not a huge file anyway.

Change-Id: Ia6ee72afe2c597b6aa38c147b5e9852a6d4a8fd3

Former-commit-id: 63d60eb3eaffa94aedf7c70b63b23219d769cc68