DevSdkUpdater can now target a single platform only
This is useful for debugging, for example playing with filters
and making sure code still compiles / tests still pass. During this
phase, its easier to work locally against a single platform, and
not having to wait for the other two platforms to download and
filter saves a fair bit of time.

Change-Id: Ifa0e91fc2b4a1524bedd2fa2e660477c896ff665

Former-commit-id: b2c25ca0c49785a0569dda19df71ed154c244928