Changes BUILD file generator to be a modifier
Adds a BUILD file parser, that parses files to an AST.

The AST allows modifications and the model rules now write to the AST
instead of a text file.

The parser and writer is compatible with buildifier, and it also buildifies
files.

The parser is not complete yet, but it's enough to parse all the BUILD files
we have. Tests in a separate CL.

Change-Id: I407d78f370291e42c3163b50c400286ec64253b8

Former-commit-id: 699daba12a2871c7f52a3f0f03b9f0abda686f47