Parallelize the IR transform pass.
This is a simple change to use Java 8 parallel streams for the transform loop.
I tried parallelizing other loops, but this is the only one that made a difference.
It sped up the Google Photos app build from 24 minutes to 11 minutes.
I had to change the code not to throw IOException, because you can't throw
checked exceptions from lambdas.

Change-Id: Ie7fe6e6a72da44c6c78c1928902efaf18898c134

Former-commit-id: 6d9b40d83564f04a2824ec38caa0176c73a13475