Add a new API method to FileCache.
The existing API method FileCache#createFile allows the client of the
cache to create a file outside the cache directory. If the cached file
exists, FileCache copies it to the given output location.

In some cases, the client may want to use the cached file directly
without copying it to another location. In this commit, we add a new API
method FileCache#createFileInCacheIfAbsent to allow the client to do
that.

However, since the returned cached output is located inside the cache
directory, the client must never write to it, and during reading, it
must ensure that another thread is not overwriting or deleting the same
cached output at the same time.

WARNING: DO NOT use this method if the returned cached output will be
annotated with Gradle's @OutputFile or @OutputDirectory as it is
undefined behavior of Gradle incremental builds when multiple Gradle
tasks have the same output-annotated file/directory.

Test: Covered by new unit tests in FileCacheTest.
Change-Id: Ib39363fe0840e0c0e1c7534ec103e388b8e96375

Former-commit-id: be95ec638508ea85a0b77cea6f2aa016e2feb7bf