Profiles: Record span threadId in ThreadRecorder.
So work is attributed to threads in profiles.

AyncThreadRecorder is replaced by the ProfileRecordWriter which
ProcessProfileWriter (was ProcessRecorder) implements directly.

Therefore the async methods can be removed from Recorder and its other
implementation ThreadRecorder, so it now has a single responsibility:
run a Recorder.Block of code and record its execution with the
ProfileRecordWriter.

Test: ProcessRecorderTest, RecordingBuildListenerTest,
PerformanceInfrastructureTest
Change-Id: I7b47fdf3f8f4c3a585d173be7d2c4c8e95ff99ea

Former-commit-id: f3ff86271270a60788ca7a45bf74864fa4d6f0dc