Force interpreted mode for IR bootstrapping.
This change uses a quasi-documented ART feature in which setting
the optimized code path to null, when creating a dex class loader,
will prevent dex2oat from running, and will instead force the class
to be interpreted.

Without this change, my Nest application failed with an ANR dialog
with IR enabled.  With the change, it starts up in under 5 seconds.

Change-Id: Ia787b752586369c9d00569461e87d13ae8d4b781

Former-commit-id: caf781f437edf49910716a5dcbe013472eb691d2