Updating behavior of simple events to match mock.
This change fixes a bug where up events didn't have an associated down
event time, as such we computed it. The change enables us to use the time
from the java system. This is far more reliable when we get an up event in our
range but not the associated down event.

Test: Ran Android Studio to manually test and verify, also ran the VisualTest
framework to verify that the behavior matched expetations.

Change-Id: Icc61e6be3211dc45d98d6f8328d0b42b868674df

Former-commit-id: 667d3a5383dc9c70af94f226d3f21035e7f66e74