Allow marking all entries as noCompress.
noCompress in aaptOptions used to allow the value '""' to
mark all files as non-compressed. This was a "hack" as this
passed -0 "" to aapt.

Since assets stopped going through aapt, this option no
longer works.

This CL fixes the base problem, the fact that gradle does
not accept empty arguments to processes. So, now we
mark noCompress '' and aapt gets -0 "".

Bug: http://b.android.com/224273
Test: Included in CL
Change-Id: Iff3539ebe7dd208ee39b31661b37de2f1ae80231

Former-commit-id: a4f573e7178b19f8bdc92e840e0e39f6cfa1b82b