Fix for aapt slaves not finished nor shutting down before task returns as complete.
Code was added to do start/compile/end at each iteration delaying the aapt slave shutdown
instead of doing start/compile*/end which it was intended to.
1. This refactoring was not properly shutting down the queued cruncher resulting in crunching
requests being outstanding.
2. To reuse the aapt process for all these compile requests, a delayed shutdown of the process
was put in place in an attempt to reuse the living processes. This in turn created the issue
where images were not entirely flushed by the aapt process/kernel.

Revived the original design based on start/compile*/end with end() waiting for all the requests
to be processed and shutdown the aapt slaves (ensuring proper buffer flushing) before returning.

Bug: 226879
Test: adapted current tests.

Change-Id: I3b5c57063a85a7b5618afef74118aee8c45eba16

Former-commit-id: a88527aa6c1e93e330329c8f4b315effd42d619a