Fix misleading error message for nested namespaces on layout tags
If you had an xmlns:android namespace declaration on a non-root layout
tags, the missing prefix detector (which looks for suspicious attributes
on layout tags, such as using "id" instead of "android:id") would kick
in and complain. That error message was misleading. Change this to be
an unused namespace warning instead.

Test: New unit test included.
Change-Id: I8f6cee4ed51db73e0cad102209659887c0afa81e

Former-commit-id: fa6a02831ac0ab474d324a51242e1508e24ad725