Add explicit revision qualifier to install/uninstall operations details
While some packages appear to have some version information in
their display names, others don't have it at all. Those which do,
may have it not fully-qualified. It seems reasonable to add
explicit revision qualifier basing on package version properties
anyway, so that it was clear which package the install/uninstall
operation is dealing with. Even if the package display name has
fully-qualified version, duplication won't hurt, but if it's
missing at all, it seems worse.

Bug: http://b.android.com/223425
Test: Added installerProperties/uninstallerProperties tests
Change-Id: Idb7e90f86e5429e9c8127076af8d0433e4890e75

Former-commit-id: 8e18b46eee0c9ea9e50c3431170da4793dc78f67