Adding better support for in-progress sdk installs
Prior to this change, if an install were initiated for a package for
which an install was already in progress, we would find the existing
PackageOperation correctly, but wouldn't try to avoid duplicating
actual operations.
After this change the UI should be able to pick up in the middle of an
ongiong operation correctly.

Also fixed a couple other errors uncovered by the test added here:
- Some synchronization problems
- It seems ordering of SchemaModules is important, and we were using a
  HashSet to store them. Switching to a List.

Fixes b.android.com/227755

Test: AbstractPackageOperationTest

Change-Id: Ieacded9e8a05131cc43e0daf088d809312dd61a5

Former-commit-id: b22d191260a55690619a8b2c67c6e8ba10d0c3c8