Start IR server via a service, and remove pre-L coldswap support
This code removes the support for dex coldswap from the runtime
library, which had a bunch of hacky code to manipulate class loaders,
transfer and extract slices from the APK, handle .dex files sitting
in an inbox directory when pushed while app wasn't running, etc.

This provides significant savings to the IR runtime: not only is it
nearly 40% smaller, it also drops the method count overhead by ~30%:
Prior to this change the IR runtime library defined
   35 classes, 182 methods, 418 method references
and after this cleanup, the counts are
   27 classes, 127 methods, 293 method references

More importantly, by dropping the pre-L coldswap support we no longer
need to patch the the user's application class, so BootstrapApplication
is gone, and we instead start the instant run server (which listens
for IDE hotswap and resource swap requests) via a system service,
InstantRunService, which is started with adb am startservice commands
from the IDE during deployment.

Test: Mostly deleted code. Instant Run live code is tested manually
by QA.

Change-Id: I8ac0dc1bfbe6cfdedd9b7cf2fed701493f6d5275

Former-commit-id: 896eef2c1b1b2e4f4ee3e7ec5d61dd68e074a789