Fix getPipelineOutputAsFileCollection.
We need to ensure we don't call FileCollection.getFiles()
when we don't have too.

The previous implementation worked for IntermediateStreams
but not for OriginalStreams as it would resolve the configuration
too early calling getFiles().

This change move the intermediate logic into the intermediate
stream and we now combine the collection together.

Test: existing tests.

Change-Id: I9ad39feda7b4955c063e4eba3af5a89c9945776d

Former-commit-id: f55c157f804325f146a733820b3882eeaa5b9ece