Pre-dexing as a separate transform
This extracts the pre-dexing as a separate
transform (c/p from DexTransform.java), while
massively simplifying DexTransfrom.

PreDexTransform - if pre-dexing libraries is enabled,
and we are mono-dex or native multidex, we will
process each .jar and get a DEX file.

DexTransform - if pre-dexing was not enabled, or we
are not in native multidex mode (we need to merge
DEX for mono-dex, or produce the legacy multidex).

Also, this will make transition to legacy multidex
pre-dexing easier.

Test: existing tests
Change-Id: Ic9f20e0d71bc594bb05a6eb1469faf2c9e335bbc

Former-commit-id: 0ed3c14c0ea255d4b5591fde5181a2a40eb94cb1