Add UTF-enforcing options to bazel.rc
This change adds a bunch of arguments to bazel.rc so they are defined
when bazel test is invoked. Specifically they are:
-Environment variable: LC_ALL=en_US.UTF-8
-JVM arg: -Dsun.jnu.encoding=UTF-8
-JVM arg: -Dfile.encoding=UTF-8

These arguments are needed to be set because without them bazel
defaults to ANSI encoding which is not sufficient as some of our tests
involve characters outside of ANSI range.

Test: Run all bazel studio tests
Change-Id: I4692ecbcdde06c4cec2150a3f49089a75c42a5a1

Former-commit-id: bfa21fbc096205405c9111c748df4d527baedfe6