Change how local jars are handled.
For inter-project aar consumption, the local (file dependencies)
jars are now not processed by the library module anymore.

This is in part due to the new dependency API showing file
dependencies transitively. Because they are present in the graph,
there is no need for the library module to republish them.

This still happens for published (to artifact repo) AARs though.

We also disasble proguard from running on library module as it
makes little sense to do it there (unless it's the AAR that's being
built), and java resource conflict resolution does not happen on
local jars (again unless you build the aar).

Test: fixing them.
Change-Id: I67937801e68c775df2d984f7fca4dcbacee779ae

Former-commit-id: f1a75325efb4ebd6c3878d7a365601a5a6a15b72